# 03-Src/Docs/openapi.yaml
openapi: 3.0.3
info:
  title: MotoStock API
  version: 1.0.0
  description: API REST versionada para importar, consolidar y consultar repuestos de moto.
servers:
  - url: http://127.0.0.1:8000/api/v1
paths:
...
  /matches/pending:
    get: {tags: [Coincidencias], parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PerPage'}], responses: {'200': {$ref: '#/components/responses/Success'}}}
  /matches/{suggestion_id}/confirm:
    post: {tags: [Coincidencias], parameters: [{name: suggestion_id, in: path, required: true, schema: {type: integer}}], responses: {'200': {$ref: '#/components/responses/Success'}, '404': {$ref: '#/components/responses/Error'}, '409': {$ref: '#/components/responses/Error'}}}
  /matches/{suggestion_id}/reject:
    post: {tags: [Coincidencias], parameters: [{name: suggestion_id, in: path, required: true, schema: {type: integer}}], responses: {'200': {$ref: '#/components/responses/Success'}, '404': {$ref: '#/components/responses/Error'}}}
...
  /agent/chat:
    post: {tags: [Agente], requestBody: {required: true, content: {application/json: {schema: {type: object, required: [message], properties: {message: {type: string}, session_id: {type: string}}}}}}, responses: {'200': {$ref: '#/components/responses/Success'}, '429': {$ref: '#/components/responses/Error'}}}
  /agent/confirm:
    post: {tags: [Agente], responses: {'200': {$ref: '#/components/responses/Success'}, '404': {$ref: '#/components/responses/Error'}}}
  /agent/actions:
    get: {tags: [Agente], parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PerPage'}], responses: {'200': {$ref: '#/components/responses/Success'}}}
...
components:
  parameters:
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PerPage: {name: per_page, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
  schemas:
...
    Envelope: {type: object, properties: {data: {}, meta: {type: object}}}
    Error: {type: object, required: [error], properties: {error: {type: object, properties: {code: {type: string}, message: {type: string}, details: {}}}}}
  responses:
    Success: {description: Operación exitosa, content: {application/json: {schema: {$ref: '#/components/schemas/Envelope'}}}}
...
    Error: {description: Error de API, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}